+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Update react-hot-loader to v4 release
+### Added
+- Plop generator for React Redux container
+- Plop generator for React components
 
+### Changed
+- Update react-hot-loader to v4 release
 
 ## 1.0.0 - 2018-03-02
 ### Added